Add texts for couples to StepSteuerlotse
